LayoutTests/imported/w3c/web-platform-tests/wai-aria/role/roles.tentative-expected.txt

Tests new ARIA role definitions. See comments for more info.

x
x

FAIL role: sectionheader assert_equals: <div role="sectionheader">x</div> expected "sectionheader" but got "generic"
FAIL role: sectionfooter assert_equals: <div role="sectionfooter">x</div> expected "sectionfooter" but got "generic"
